+++ packages/suite/src/components/suite/Preloader/__tests__/Preloader.test.tsx
@@ -59,6 +59,7 @@ export const getInitialState = ({ suite, router }: any = {}) => ({
             banner: [],
             context: [],
             modal: [],
+            feature: [],
         },
     },
     modal: {